requirements.txt
pandas>=2.0
pyarrow>=14.0
polars>=0.20
psutil>=5.9
lxml>=4.9
rapidfuzz>=3.0
pyyaml>=6.0
openpyxl>=3.1
xlsxwriter>=3.1
